fix(quartodoc): search up through `BaseBackend` for docstrings
We were looking at the `resolved_bases` to find docstrings that were
defined in a parent class, but this search only goes up a single level
of inheritance.  For all of our sql-based backends, this means we only
go up to `SQLBackend`, not `BaseBackend`.

Manually adding `BaseBackend` to the resolution list fills out a LOT of
missing docstrings on the backend docs pages.