chore(repr): simplify Expr.__rich_console__(), move generic error handling up to _repr_mimebundle_
First, remove the unneeded usage of rich.Text. Since we aren't applying any styles, wrapping the vanilla python str in rich.Text has no effect.

Second, move the generic exception catching and printing to the _repr_mimebundle_() method.
If an exception occurs during plain usage of __rich_console__, not in the context of ipython, then the error will get surfaced as normal. Only if we are in the context of ipython would the error get sliently eaten. Therefore, only deal with this edge case in _repr_mimebundle().
We now print the exception with traceback.print_exc(), this might be slightly different from console.print_exception(), but similar enough.

Third, simply pass the options.max_width to pretty.to_rich(). This causes, when in jupyter, this function to be called with a max_width of 1_000_000 instead of None (which would get interpreted as infinite). In functionality, these both mean the same thing since 1_000_000 is effectively infinite, but this new method uses slightly less branching logic.