feat(decompile): make the decompiler run on TPCH query 1 (#9779)
Context: the sql -> ibis expression parser/creator runs off of a
`sqlglot.Plan` object

The `sqlglot` plan optimizer replaces a few items with aliases. In other
places, it doesn't use
fully-qualified column names. Both of these prevent us from reliably
converting into an Ibis expression.

## Description of changes

### Dereferencing aggregation operands

The sqlglot planner will pull out computed operands into a separate
section and alias them e.g.
```
Context:
    Aggregations:
    - SUM("_a_0") AS "sum_disc_price"
    Operands:
    - "lineitem"."l_extendedprice" * (1 - "lineitem"."l_discount") AS _a_0
```

For the purposes of decompiling, we want these to be inline, so here we
replace those new aliases with the parsed sqlglot expression

### Dereferencing projections

The sqlglot planner will (sometimes) alias projections to the aggregate
that precedes it.

```
- Sort: lineitem (132849388268768)
  Context:
    Key:
      - "l_returnflag"
      - "l_linestatus"
  Projections:
    - lineitem._g0 AS "l_returnflag"
    - lineitem._g1 AS "l_linestatus"
    <snip>
  Dependencies:
  - Aggregate: lineitem (132849388268864)
    Context:
      Aggregations:
        <snip>
      Group:
        - "lineitem"."l_returnflag"  <-- this is _g0
        - "lineitem"."l_linestatus"  <-- this is _g1
        <snip>
```

These aliases are stored in a dictionary in the aggregate `groups`, so
if
those are pulled out beforehand then we can use them to replace the
aliases in the projections.

### Dereferencing sort keys

The sqlglot planner doesn't fully qualify sort keys

```
- Sort: lineitem (132849388268768)
  Context:
    Key:
      - "l_returnflag"
      - "l_linestatus"
```

~For now we do a naive thing here and prepend the name of the sort~
~operation itself, which (maybe?) is the name of the parent table.~
~*This is definitely the most brittle part of the existing decompiler*~
 
I've mucked around with this a little bit, and while it's a _touch_
hacky, using the deferred operator with the not-fully-referenced
sort-keys works well.