fix(duckdb): fix create_table() in databases with spaces in the name (#9817)
If you tried

```python
import ibis

conn = ibis.duckdb.connect("with space.duckdb")
conn.create_table("foo", {"a": [0,1,2]})
```

then you would get the DDL `CREATE TABLE with space.main."foo" ("a"
BIGINT)`. Now, the database name is quoted.

I went through the create_table() method in all the backends and looked
for where we forgot to use "quoted", but I almost definitely missed some
cases.

I need to add a test for this, but I would love your help with this:
- should it add/modify a create_table() test to check for a database
with spaces in the name?
- should it go somewhere near other tests that check for spaces? I think
we have those?
- or should we go larger, and modify the `con` fixture to sometimes be a
database with spaces in the name?