docs(algolia): also add functions to algolia index
I wrote some hacky stuff a while back to add the various methods to the
Algolia index so that they would show up in search.

This is basically the same thing again, except it grabs the top-level
anchored functions that are generated by `quartodoc` (and which are not
included in the quarto-generated `search.json` file).

I've run this already to populate the index manually, and now functions
like `ibis.get_backend` and `ibis.set_backend` now show up in search.