fix(insert): user can specify insert table in specified database
A user reported this issue on Zulip.

We weren't passing `catalog.database` information to `get_schema` inside
of `_build_insert_from_table` which caused inserts to fail if user had
specified a `database` argument, because we were looking in the wrong db
for the existing table to pull the `schema` from.