docs(examples): add `having` example for GroupedTable (#10457)
<!--
Thanks for taking the time to contribute to Ibis!

Please ensure that your pull request title matches the conventional
commits
specification: https://www.conventionalcommits.org/en/v1.0.0/
-->

## Description of changes

This adds a
[`having`](https://ibis-project.org/reference/expression-tables#ibis.expr.types.groupby.GroupedTable.having)
example for usage on a `GroupedTable`. I followed it up with an
`aggregation` to show the removal of the "a" rows.

It could be useful also to show the `having` method without aggregation
results in a `GroupedTable,` but I wasn't 100% sure if that would be
worth showing.