docs: reword backend counting to nearly 20 from 20+ (#10533)
## Description of changes

Back in https://github.com/ibis-project/ibis/pull/10259, I adjusted the
messaging around the backend count in the README to represent the
changes after removing the pandas and Dask backends. This PR applies
this fix in more locations. I did not adjust any presentations or
previous blog posts, as the count was correct then.