refactor(impala): remove impala table (#9840)
This removes the `ImpalaTable` subclass. Relevant methods that were on
that class have been moved to the `Backend` class instead.

We used to have backend-specific subclasses of `Table` that added some
backend-specific methods. Most of those were removed in prior releases,
`ImpalaTable` was the last remaining one.

As this is a breaking change, we should hold off on merging this until
we're ready to start prepping for a 10.0 release. Depends on #9839.

Fixes #9365.

---------

Co-authored-by: Gil Forsyth <gil@forsyth.dev>