refactor(join): change default lname from "" to "{name}"
There is this code inside expr/types/join.py:

```python
def disambiguate_fields(
    how,
    predicates,
    equalities,
    left_fields,
    right_fields,
    left_template,
    right_template,
):
    """Resolve name collisions between the left and right tables."""
    collisions = set()
    left_template = left_template or "{name}"
    right_template = right_template or "{name}"
   ...
```

This makes it so that "" is treated equivalently to "{name}". So this change doesn't have any behavior change, it just makes the defaults follow a more consistent pattern. The old patterns was a remnant from the lsuffix API we used to have.