docs: add introduction to pandas users for adding columns with single text values (literal) (#11010)
## Description of changes

ibis generally follows pandas syntax for mutating columns, with the
exception of adding string columns where `ibis.literal` is needed.
`ibis.literal` is not mentioned yet in the pandas' user introductions,
so it would be helpful to include this as it's not clear to the user why
adding a float column or summing two columns together uses the same
syntax as pandas, while a adding a single-valued string column does not.

This is the reason
https://github.com/ibis-project/ibis/discussions/11009 was opened (as a
q&a link to send to new users).

Including it in the pandas introduction docs would reduce friction for
switching to ibis for pandas users.