feat(pyspark)!: forward kwargs in `create_table` to respective pyspark methods
Forward kwargs in `create_table` method to respective pypsark methods.
Also, we removed `partition_by` and `format` kwargs of `create_table` in favor
of forwarding kwargs directly to pyspark methods. This is consistent
with `to_delta` and `to_parquet` for pyspark backend.

This is breaking for users of `partition_by` kwarg only in `create_table`,
which will now need to be passed as `partitionBy`.

closes #10984