feat(duckdb): don't translate ops.First() to ops.FirstValue() in windows
Do not replace first() with first_value() because, per
https://duckdb.org/docs/stable/sql/functions/window_functions#aggregate-window-functions
The first and last aggregate functions are shadowed by the respective
general-purpose window functions, with the minor consequence that the
FILTER clause is not available for these but IGNORE NULLS is.

After this change, the generated SQL is more canonical, and include_null actually works.

This probably should wait until https://github.com/ibis-project/ibis/pull/11311 lands, which has the tests for this change.