feat: support dtype(ir.Value) and ir.Value[<dtype>]
This allows for pleasant UX such as
- ibis.dtype(ibis.ir.StringValue)
- ibis.dtype(ibis.ir.IntegerValue["!uint32"])

```python
class Users(ibis.Table):
    name: ibis.ir.StringColumn
    age: ibis.ir.IntegerColumn["uint8"]

def my_api(users: Users):
    # Runtime coercion of types works as expected!
    users = ibis.cast(users, Users)
    # IDE type hints work too!
    reveal_type(users.age) # Its an IntegerColumn!
```

I also adjusted the ibis.dtype() function so that the nullable kwarg defaults to None, which means "Don't mess with the nullability of something that is already a dtype". I can put this into a separate PR if desired, but I think this is a good change to make anyways.
The current behavior of `dtype(nonnullable_dtype, nullable=True)` returning the original input, still nonnullable, is a footgun.